Add getAllQuizzes functionand basic page layout
