+++ views/home-page.ejs
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+  <link rel="stylesheet" href="/vendor/normalize-4.1.1.css" />
+  <link rel="stylesheet" href="/vendor/border-box.css" />
+  <link rel="stylesheet" href="/styles/main.css" />
+  <link rel="stylesheet" href="/styles/layout.css" />
+  <link rel="stylesheet" href="/styles/header.css" />
+
+
+  <script src="/vendor/jquery-3.0.0.js"></script>
+  <script defer src="/scripts/app.js"></script>
+  <title>Home Page</title>
+</head>
+<body>
+  <%- include('partials/_header') %>
+
+  <main>
+    <h3>Quizzes</h3>
+
+    <div class="card">
+      <div class="card-body">
+        <% for (let quiz of quizzes) { %>
+        <h5 class="card-title"><%= quiz.title %></h5>
+        <p class="card-text"><%= quiz.description %></p>
+        <a href="#" class="btn btn-primary">Go somewhere</a>
+        <% } %>
+      </div>
+    </div>
+
+  </main>
+</body>
+</html>
+
+
