+++ views/partials/_header.ejs
@@ -3,8 +3,8 @@
     <div class="nav-container">
       <span>Quizhoot</span>
       <button href="/login"> Login </button>
-      <button href="/login"> Home </button>
-      <button><a href="quizzes/new">Create a Quiz</a></button>
+      <button><a href="/quizzes/home"> Home</a></button>
+      <button><a href="/quizzes/new">Create a Quiz</a></button>
       <button href="quizzes/new"> Profile </button>
       <button href="/login"> Logout </button>
